Former-commit-id: 4ac5e3a850f5ac62a4426d78296d90f6b8c344f3

+++ Arduino/Sketchbook/libraries/ros_lib/arduino_msgs/ArduinoFeedback.h
@@ -8,7 +8,6 @@
 #include "std_msgs/Header.h"
 #include "std_msgs/Float32.h"
 #include "std_msgs/Bool.h"
-#include "std_msgs/Int8.h"
 #include "std_msgs/Int16.h"
 
 namespace arduino_msgs
@@ -23,26 +22,26 @@ namespace arduino_msgs
       std_msgs::Bool LF_motor_enable;
       std_msgs::Bool LF_motor_dir;
       std_msgs::Float32 LF_wheel_rpm;
-      std_msgs::Int8 LF_motor_cmd;
-      std_msgs::Int8 LF_servo_angle;
+      std_msgs::Int16 LF_motor_cmd;
+      std_msgs::Float32 LF_servo_angle;
       std_msgs::Int16 LF_servo_cmd;
       std_msgs::Bool RF_motor_enable;
       std_msgs::Bool RF_motor_dir;
       std_msgs::Float32 RF_wheel_rpm;
-      std_msgs::Int8 RF_motor_cmd;
-      std_msgs::Int8 RF_servo_angle;
+      std_msgs::Int16 RF_motor_cmd;
+      std_msgs::Float32 RF_servo_angle;
       std_msgs::Int16 RF_servo_cmd;
       std_msgs::Bool LR_motor_enable;
       std_msgs::Bool LR_motor_dir;
       std_msgs::Float32 LR_wheel_rpm;
-      std_msgs::Int8 LR_motor_cmd;
-      std_msgs::Int8 LR_servo_angle;
+      std_msgs::Int16 LR_motor_cmd;
+      std_msgs::Float32 LR_servo_angle;
       std_msgs::Int16 LR_servo_cmd;
       std_msgs::Bool RR_motor_enable;
       std_msgs::Bool RR_motor_dir;
       std_msgs::Float32 RR_wheel_rpm;
-      std_msgs::Int8 RR_motor_cmd;
-      std_msgs::Int8 RR_servo_angle;
+      std_msgs::Int16 RR_motor_cmd;
+      std_msgs::Float32 RR_servo_angle;
       std_msgs::Int16 RR_servo_cmd;
 
     virtual int serialize(unsigned char *outbuffer)
